% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{augment.ipf_rake}
\alias{augment.ipf_rake}
\title{Augment data with raked weights}
\usage{
\method{augment}{ipf_rake}(x, ...)
}
\arguments{
\item{x}{An \code{ipf_rake} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A tibble with all original columns plus \code{.weight}.
}
\description{
Returns the original data frame with a \code{.weight} column appended.
}
\examples{
data <- data.frame(
  gender = sample(c('M', 'F'), 100, replace = TRUE, prob = c(0.6, 0.4))
)
targets <- list(gender = c(M = 0.5, F = 0.5))
result <- rake(data, targets)
augment(result)

}
