% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{design_effect_rust}
\alias{design_effect_rust}
\title{Compute design effect and effective sample size from a weight vector.}
\usage{
design_effect_rust(weights)
}
\arguments{
\item{weights}{Numeric weight vector.}
}
\description{
Compute design effect and effective sample size from a weight vector.
}
