% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf-package.R
\docType{package}
\name{ipf-package}
\alias{ipf}
\alias{ipf-package}
\title{ipf: Iterative Proportional Fitting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fast raking for survey weighting. The computational core is written in Rust for speed. Supports multiple raking variables, automatic variable selection, weight bounding, and comprehensive diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://christophertkenny.com/ipf/}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
