% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{weight_summary_rust}
\alias{weight_summary_rust}
\title{Compute summary statistics for a weight vector.}
\usage{
weight_summary_rust(weights)
}
\arguments{
\item{weights}{Numeric weight vector.}
}
\description{
Compute summary statistics for a weight vector.
}
