% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf-package.R
\docType{package}
\name{ipf-package}
\alias{ipf}
\alias{ipf-package}
\title{ipf: Iterative Proportional Fitting (Raking) with a Rust Backend}
\description{
Fast iterative proportional fitting (raking) for survey weighting. The computational core is written in Rust for speed and numerical stability. Supports multiple raking variables, automatic variable selection, weight bounding, and comprehensive diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://christophertkenny.com/ipf/}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
