% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf_rake-class.R
\name{print.ipf_rake}
\alias{print.ipf_rake}
\title{Print an ipf_rake object}
\usage{
\method{print}{ipf_rake}(x, ...)
}
\arguments{
\item{x}{An \code{ipf_rake} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ipf_rake object
}
